+++ CodeMover/Control/Strategies/DestinationStrategy.cs
@@ -9,11 +9,13 @@ namespace CodeMover.Control.Strategies
 {
    public class DestinationStrategy : IStrategy, IArgs
    {
+      #region Properties
       public string Args { get; set; }
       public Status Status { get; set; }
       public string Results { get; set; }
-      //object IStrategyResults.Results { get => Results; set => Results = value as string; }
+      #endregion
 
+      #region Methods
       public async Task<Status> Run()
       {
          Status = Status.working;
@@ -27,5 +29,6 @@ namespace CodeMover.Control.Strategies
             return Status.done;
          });
       }
+      #endregion
    }
 }
