+++ CodeMover/Control/Strategies/ExecuteStrategy.cs
@@ -10,16 +10,17 @@ namespace CodeMover.Control.Strategies
 {
    public class ExecuteStrategy : IStrategy, IFileResults
    {
+      #region Properties
       public Status Status { get; set; }
       public List<FileRecord> Results { get; set; }
-      //object IStrategyResults.Results { get => Results; set => Results = value as List<FileRecord>; }
+      #endregion
 
+      #region Methods
       public async Task<Status> Run()
       {
          Status = Status.working;
          try
          {
-            //Results = await FileController.MoveDirAsync();
             Results = await FileController.NewMoveDirAsync();
          }
          catch (Exception)
@@ -34,5 +35,6 @@ namespace CodeMover.Control.Strategies
          Results = FileController.MoveDir().ToList();
          return Status.done;
       }
+      #endregion
    }
 }
