+++ CodeMover/Control/Strategies/ExecutionTestStrat.cs
@@ -9,10 +9,12 @@ namespace CodeMover.Control.Strategies
 {
    public class ExecutionTestStrat : IStrategy, IPathResult
    {
+      #region Properties
       public Status Status { get; set; }
       public string Results { get; set; }
-      //object IStrategyResults.Results { get => Results; set => Results = value as string; }
+      #endregion
 
+      #region Methods
       public async Task<Status> Run()
       {
          Status = Status.working;
@@ -22,5 +24,6 @@ namespace CodeMover.Control.Strategies
             return Status.done;
          });
       }
+      #endregion
    }
 }
