+++ CodeMover/Control/Strategies/ResultInterfaces/IArgs.cs
