+++ CodeMover/Logic/Copy.cs
@@ -2,86 +2,42 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Text;
-using System.Threading.Tasks;
 
 namespace CodeMover.Logic
 {
-   public class Copy : IAsyncDisposable
+   public class Copy : IDisposable
    {
       #region - Fields & Properties
-      public FileRecord File { get; set; }
-      private StreamWriter Writer { get; set; }
-      private StreamReader Reader { get; set; }
+      public FileInfo SourceFile { get; set; }
+      public FileRecord Record { get; set; }
       #endregion
 
       #region - Constructors
-      public Copy() { }
+      public Copy(FileRecord record)
+      {
+         Record = record;
+      }
       #endregion
 
       #region - Methods
-      public virtual async Task<FileRecord> CopyFileAsyncNoGC()
-      {
-         var fileData = await Task.Run(() =>
-         {
-            using (Reader = new StreamReader(File.Source))
-            {
-               return Reader.ReadToEnd();
-            }
-         });
-         return await Task.Run(() =>
-         {
-            using (Writer = new StreamWriter(File.Destination))
-            {
-               try
-               {
-                  Writer.Write(fileData);
-               }
-               catch (Exception e)
-               {
-                  File.Error = e;
-               }
-            }
-            return File;
-         });
-      }
-
       public FileRecord CopyFile()
-      {
-         Reader = new StreamReader(File.SourceFilePath);
-         var fileData = Reader.ReadToEnd();
-         Reader.Close();
-         Writer = new StreamWriter(File.DestFilePath);
-         Writer.Write(fileData);
-         Writer.Flush();
-         Writer.Close();
-         return File;
-      }
-
-      public virtual async Task<FileRecord> CopyFileAsyncNoUsing()
       {
          try
          {
-            Reader = new StreamReader(File.DestFilePath);
-            Writer = new StreamWriter(File.Destination);
-            System.IO.File.Create(File.DestFilePath);
-            var fileData = await Reader.ReadToEndAsync();
-            await Writer.WriteAsync(fileData);
+            SourceFile = new FileInfo(Record.SourceFilePath);
+            _ = SourceFile.CopyTo(Record.DestFilePath, true);
+            return Record;
          }
-         catch (Exception e)
+         catch (Exception)
          {
-            File.Error = e;
+            throw;
          }
-         return File;
       }
 
-      public ValueTask DisposeAsync()
+      public void Dispose()
       {
-         throw new NotImplementedException(nameof(DisposeAsync));
-      }
-
-      private ValueTask DisposeAsyncCore()
-      {
-         throw new NotImplementedException(nameof(DisposeAsyncCore));
+         Record = null;
+         SourceFile = null;
       }
       #endregion
